First commit on the map editor
