+++ docs/todo/There's Levels to This.md
@@ -1,0 +1,29 @@
+# There's Levels to This
+We have levels, we have stories, we have rules, we have facts.
+
+But facts are not saved, not rules either, and I'm not entirely entirely sure about the "levels as pictures"-stuff either.
+
+We should probably move ALL of this into a mapdefinition that can be serialized - and therefore loaded and unloaded, edited in a regular text editor. It **would** be cool if we could make the actual map into a simple textformat thing.
+## Example map def
+Space character is empty? 
+Or if it is empty, replace it with a star or dot or something?
+~~~
+*s*************************
+***************************
+***************************
+***************************
+***************************
+***************************
+***************************
+***************************
+***************************
+***************************
+***************************
+~~~
+
+
+```
+ecwee code boc
+```
+
+Yes
